39 fix routine name memory on creation (#42)
* added Cancel to SelectFolderModal as well as replaced FolderToggle with a simple mapped list of folder names

* Routine Name is now preserved when entered before adding exercises to routine, also preserved if add exercise is cancelled.

* fixed notes in ExceriseRoutineCard to be a text area instead of input

* lots of changes to the schema, and tweaking a number of files so that we now Start Routine and it routes to its own session based on the auth user.
feels like a cluster fuck right now, def gonna need some refactoring once this thing is up and running.
getting close. I hope.

* session now wrapped in layout with timer and finish button
logic to come next

* exercises now load with previous session data.
it took a while to figure out why it wasn't rendering the data - had to pass it to the ExceriseRoutineCard.
but i was also dealing with an infinite loop issue on the SessionClient. sorted it too.

* wired up resttimer functionality

* styled checkbox component